Remove unused DU 1.0 related code
