+++ app/src/marketplace/containers/ProductPage/MembersGraph.jsx
@@ -1,93 +1,76 @@
 // @flow
 
-import React, { useEffect, useState, useCallback, useRef, useMemo } from 'react'
-import { useSubscription } from 'streamr-client-react'
+import React, { useEffect, useState, useMemo, useRef, useCallback } from 'react'
 
+import useIsMounted from '$shared/hooks/useIsMounted'
 import ClientProvider from '$shared/components/StreamrClientProvider'
 import TimeSeriesGraph from '$shared/components/TimeSeriesGraph'
-import useIsMounted from '$shared/hooks/useIsMounted'
+import { getJoinsAndParts } from '$mp/modules/dataUnion/services'
 
 type Props = {
-    joinPartStreamId: ?string,
+    dataUnionAddress: string,
     memberCount: number,
     shownDays?: number,
 }
 
-type JoinPartMessage = {
-    type: string,
-    addresses: Array<string>,
-}
-
-type MessageMetadata = {
-    messageId: {
-        timestamp: number,
-    }
-}
-
 const MILLISECONDS_IN_DAY = 24 * 60 * 60 * 1000
 
-const MembersGraph = ({ joinPartStreamId, memberCount, shownDays = 7 }: Props) => {
+const MembersGraph = ({ dataUnionAddress, memberCount, shownDays = 7 }: Props) => {
     const isMounted = useIsMounted()
     const [memberCountUpdatedAt, setMemberCountUpdatedAt] = useState(Date.now())
     const [memberData, setMemberData] = useState([])
     const [graphData, setGraphData] = useState([])
-    const activeAddressesRef = useRef([])
+    const generator = useRef(null)
 
     const startDate = useMemo(() => (
         Date.now() - (shownDays * MILLISECONDS_IN_DAY)
     ), [shownDays])
 
-    const onMessage = useCallback((data: JoinPartMessage, metadata: MessageMetadata) => {
-        if (!isMounted()) { return }
+    const reset = useCallback(() => {
+        setGraphData([])
+        setMemberData([])
+    }, [])
 
-        let diff = 0
-        const activeAddresses = activeAddressesRef.current
-        let msgAddresses = data.addresses
+    useEffect(() => {
+        setMemberCountUpdatedAt(Date.now())
+    }, [memberCount])
 
-        // Sometimes data.addresses is not an array but raw string instead.
-        // Correct this to be an array.
-        if ((data.type === 'join' || data.type === 'part') && !Array.isArray(msgAddresses)) {
-            msgAddresses = [((data.addresses: any): string)]
-        }
-
-        // Check if message type is 'join' or 'part' and
-        // calculate member count diff based on the type.
-        // E.g. 'join' with 3 addresses means +3 diff in member count.
-        // JoinPartStream might have duplicate joins/parts for a
-        // single address so make sure we skip duplicates.
-        if (data.type === 'join' && msgAddresses && msgAddresses.length > 0) {
-            msgAddresses.forEach((address) => {
-                if (!activeAddresses.includes(address)) {
-                    diff += 1
-                    activeAddresses.push(address)
+    useEffect(() => {
+        const loadData = async () => {
+            try {
+                if (generator.current != null) {
+                    generator.current.return('Canceled')
+                    generator.current = null
+                    reset()
                 }
-            })
-        } else if (data.type === 'part' && msgAddresses && msgAddresses.length > 0) {
-            msgAddresses.forEach((address) => {
-                if (activeAddresses.includes(address)) {
-                    diff -= 1
-                    const addrIndex = activeAddresses.indexOf(address)
-                    if (addrIndex > -1) {
-                        activeAddresses.splice(addrIndex, 1)
+                generator.current = getJoinsAndParts(dataUnionAddress, startDate)
+
+                // eslint-disable-next-line no-restricted-syntax
+                for await (const event of generator.current) {
+                    if (isMounted()) {
+                        setMemberData((prev) => [
+                            ...prev,
+                            event,
+                        ])
                     }
                 }
-            })
-        } else {
-            // Reject other message types.
-            return
+            } catch (e) {
+                console.warn(e)
+            }
         }
 
-        if (diff !== 0) {
-            const entry = {
-                timestamp: metadata.messageId.timestamp,
-                diff,
-            }
-            setMemberData((oldArray) => [
-                ...oldArray,
-                entry,
-            ])
+        if (dataUnionAddress) {
+            loadData()
         }
-    }, [isMounted])
+    }, [dataUnionAddress, startDate, reset, isMounted])
+
+    useEffect(() => () => {
+        // Cancel generator on unmount
+        if (generator.current != null) {
+            generator.current.return('Canceled')
+            generator.current = null
+        }
+    }, [])
 
     useEffect(() => {
         memberData.sort((a, b) => b.timestamp - a.timestamp)
@@ -99,7 +82,7 @@ const MembersGraph = ({ joinPartStreamId, memberCount, shownDays = 7 }: Props) =
             y: memberCount,
         }]
 
-        // Because we cannot read the whole joinPartStream, we have to
+        // Because we cannot read every event from blockchain, we have to
         // work backwards from the initial state and calculate graph points
         // using the member count diff.
         const data = memberData.reduce((acc, element) => {
@@ -127,29 +110,6 @@ const MembersGraph = ({ joinPartStreamId, memberCount, shownDays = 7 }: Props) =
         })
         setGraphData(data)
     }, [memberData, memberCount, memberCountUpdatedAt, startDate])
-
-    useEffect(() => {
-        setMemberCountUpdatedAt(Date.now())
-    }, [memberCount])
-
-    useEffect(() => {
-        // Clear member data when we change shownDays because
-        // resubscription to stream will happen and data will
-        // be resent
-        setMemberData([])
-        activeAddressesRef.current = []
-    }, [shownDays, joinPartStreamId, onMessage])
-
-    useSubscription({
-        stream: joinPartStreamId,
-        resend: {
-            from: {
-                timestamp: startDate,
-            },
-        },
-    }, {
-        onMessage,
-    })
 
     return (
         <TimeSeriesGraph
