+++ app/src/marketplace/modules/product/services.js
@@ -13,7 +13,6 @@ import type { StreamList } from '$shared/flowtype/stream-types'
 import { getValidId, mapProductFromApi, mapProductToPostApi, mapProductToPutApi } from '$mp/utils/product'
 import { getProductFromContract } from '$mp/modules/contractProduct/services'
 import { fromAtto, toAtto } from '$mp/utils/math'
-import { getDefaultDataUnionVersion } from '$mp/modules/dataUnion/services'
 import { productTypes } from '$mp/utils/constants'
 import routes from '$routes'
 import { getContract, call, send } from '../../utils/smartContract'
@@ -72,7 +71,7 @@ export const postEmptyProduct = (type: ProductType): ApiResult<Product> => {
     if (type === productTypes.DATAUNION) {
         product = {
             ...product,
-            dataUnionVersion: getDefaultDataUnionVersion(),
+            dataUnionVersion: 2,
         }
     }
